Improved AsTerm class
The AsTerm class as implemented was too expensive in practice:
removing unused let bindings is more effectively handled by using
a dead code elimination pass after partialEval.

When checking terms for inlining / application, it is often not
necessary to force any substitutions which have not yet occurred.
As such, AsTerm now has an unsafeAsTerm which effectively discards
substitutions that have not been performed. This is *not* suitable
for normal use as it can introduce free variables.