Skeleton for implementing primitives
Primitives may want to use eval on arguments in their rules, so
a boot file is added for exposing eval. This means we can still
have primtives implemented in different modules as opposed to
throwing everything into the Eval module, which is already large.

Exceptions which primitives may throw (Alternative.empty and
EvalException) are caught in the new evalPrimitive and handled
appropriately. We can't handle arithmetic excpetions or xexceptions
here unfortunately.